poison_methods/GARAG/requirements.txt
about-time==4.2.1
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
airportsdata==20241001
alive-progress==3.2.0
annotated-types==0.7.0
anyio==4.7.0
anytree==2.12.1
astor==0.8.1
async-timeout==5.0.1
attrs==24.3.0
autograd==1.7.0
beautifulsoup4==4.12.3
beir==2.0.0
bert-score==0.3.13
bioc==2.1
blake3==1.0.0
boto3==1.35.84
botocore==1.35.84
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cma==3.2.2
compressed-tensors==0.8.1
conllu==4.5.3
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
deprecated==1.2.15
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
elasticsearch==7.9.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
flair==0.14.0
fonttools==4.55.3
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.9.0
ftfy==6.3.1
gdown==5.2.0
gguf==0.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gradio==5.9.1
gradio-client==1.5.2
grapheme==0.6.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
interegular==0.3.3
intervaltree==3.1.0
jieba==0.42.1
jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langdetect==1.0.9
language-tool-python==2.8.1
lark==1.2.2
latex2mathml==3.77.0
lemminflect==0.2.3
lightning==2.4.0
lightning-utilities==0.11.9
lm-format-enforcer==0.10.9
lru-dict==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.2
markupsafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mistral-common==1.5.1
more-itertools==10.5.0
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
nltk==3.9.1
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.58.1
opencv-python-headless==4.10.0.84
openhownet==2.0
orjson==3.10.12
outlines==0.1.11
outlines-core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
peft==0.14.0
pillow==10.4.0
pinyin==0.4.0
pluggy==1.5.0
pptree==3.1
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
prompt-toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.4
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.18.0
pymoo==0.6.1.3
pyparsing==3.2.0
pysocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.4.0
pytorch-revgrad==0.2.0
pytrec-eval==0.5
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.3
s3transfer==0.10.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
segtok==1.5.11
semantic-version==2.10.0
semver==3.0.2
sentence-transform
